fix: address Phase B code review — dangling mass, unwrap, determinism, config wiring
C1: Weighted PageRank now accumulates dangling node mass (zero out-weight
    nodes) and redistributes it uniformly, preserving the stochastic
    matrix property. Mirrors the unweighted version's pattern.

C2: Replace 3 bare unwrap() calls with expect() providing context:
    - rank map lookup in weighted PageRank
    - edge_endpoints in community detection (2 sites)

I2: Louvain comm_weights changed from HashMap to BTreeMap for
    deterministic community evaluation order. Added smallest-ID
    tiebreaker when delta values are equal.

I3: Remove redundant `as usize` cast on min_percentile_drift (already usize).

I1: Wire CoChangeConfigToml from project config to git_enrich_stage
    via new co_change_config field on GitEnrichInput. Added
    CoChangeConfigToml::to_runtime() conversion method.

+++ repotoire-cli/src/detectors/pagerank_drift.rs
@@ -154,7 +154,7 @@ impl Detector for PageRankDriftDetector {
 
             let drift = (static_pct as isize - weighted_pct as isize).unsigned_abs();
 
-            if drift <= self.min_percentile_drift as usize {
+            if drift <= self.min_percentile_drift {
                 continue;
             }
